# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API de gestion",
      "version": "1.0.0",
      "description": "API pour gérer les utilisateurs, référentiels, promotions et apprenants"
    },
    "servers": [
      {
        "url": "http://localhost:5000/api",
        "description": "Serveur Local"
      }
    ],
    "paths": {
...
      "/v1/users": {
        "get": {
          "tags": ["Users"],
          "summary": "Liste des utilisateurs",
          "operationId": "getUsers",
          "responses": {
            "200": {
              "description": "Liste des utilisateurs retournée"
            }
          }
        },
        "post": {
          "tags": ["Users"],
          "summary": "Création d'un utilisateur",
          "operationId": "createUser",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRequest"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Utilisateur créé"
            }
          }
        }
      },
...
      "/v1/referentiels": {
        "get": {
          "tags": ["Referentiels"],
          "summary": "Liste des référentiels",
          "operationId": "getReferentiels",
          "responses": {
            "200": {
              "description": "Liste des référentiels retournée"
            }
          }
        },
        "post": {
          "tags": ["Referentiels"],
          "summary": "Création d'un référentiel",
          "operationId": "createReferentiel",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferentielRequest"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Référentiel créé"
            }
          }
        }
      },
...
      "/v1/apprenants": {
        "get": {
          "tags": ["Apprenants"],
          "summary": "Liste des apprenants",
          "operationId": "getApprenants",
          "responses": {
            "200": {
              "description": "Liste des apprenants retournée"
            }
          }
        },
        "post": {
          "tags": ["Apprenants"],
          "summary": "Création d'un apprenant",
          "operationId": "createApprenant",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprenantRequest"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Apprenant créé"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "UserRequest": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "email": {
              "type": "string",
              "format": "email"
            },
            "role": {
              "type": "string",
              "enum": ["Admin", "User"]
            }
          },
          "required": ["name", "email", "role"]
        },
...
        "ReferentielRequest": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            },
            "description": {
              "type": "string"
            }
          },
          "required": ["title"]
        },
...
        "ApprenantRequest": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "email": {
              "type": "string",
              "format": "email"
            }
          },
          "required": ["name", "email"]
        }
      }
    }
  }